Refactored ORM model utils module unit testing
* Reorganised unit test cases that belong to ORM data model
  `utils` module into the `test_utils`
